refactor(app): extract EntryRoutes from App.tsx (#316 phase 8a)
The screens a session starts from: intro, title (with its FeedbackModal and
LoginModal), settings, quickbattle, carddraft, dailychallenge,
weeklychallenge, commander and training.

AppContextValue gains newsUnreadCount, feedbackOpen, showTitleLoginModal and
the entry-point handlers (handlePlay, handleDraftComplete, the two challenge
start handlers, handleStartTraining, handleResetGame, checkForUpdates).

App.tsx: 3,118 -> 2,995 lines. Build green, chunk boundaries unchanged,
2107 tests passing.
